do quest 1st time
