+++ package.json
@@ -1,12 +1,13 @@
 {
   "name": "mesto",
-  "homepage":"https://github.com/742th/mesto",
+  "homepage": "https://742th.github.io/mesto/",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
   "scripts": {
-    "build": "rimraf dist && webpack --mode production",
-    "dev": "webpack-dev-server --mode development --open --watch"
+    "build": "cross-env NODE_ENV=rimraf dist && webpack --mode production",
+    "dev": "cross-env NODE_ENV=webpack-dev-server --mode development --open --watch",
+    "deploy": "cross-env NODE_ENV=gh-pages -d dist"
   },
   "repository": {
     "type": "git",
@@ -17,7 +18,6 @@
   "bugs": {
     "url": "https://github.com/742th/mesto/issues"
   },
-  "homepage": "https://github.com/742th/mesto#readme",
   "devDependencies": {
     "@babel/cli": "^7.8.4",
     "@babel/core": "^7.8.7",
@@ -29,10 +29,13 @@
     "css-loader": "^3.4.2",
     "cssnano": "^4.1.10",
     "file-loader": "^6.0.0",
+    "gh-pages": "~2.2.0",
     "html-webpack-plugin": "^3.2.0",
     "image-webpack-loader": "^6.0.0",
     "mini-css-extract-plugin": "^0.9.0",
+    "optimize-css-assets-webpack-plugin": "^5.0.3",
     "postcss-loader": "^3.0.0",
+    "style-loader": "^1.1.3",
     "webpack": "^4.42.0",
     "webpack-cli": "^3.3.11",
     "webpack-dev-server": "^3.10.3",
