+++ scripts/api.js
@@ -1,13 +1,8 @@
-//спасибо за терпение, и что все так объяснили, я реально путаюсь с этими кетч и классами еще, трудновато понять как все это должно работать,
-// и трудно еще привыкнуть к последовательности работы с сервером, что сначала хапрос нужно потом то что вернулось уже как-то прменить и тд + асинхрон, в общем еще раз спасибо,
-// я не специально такой тупой
-// а это я еще доп задания не успел сделать :(
-
-// Думаю что как раз лучше попробовать доп. задания сделать, чтобы руку набить. Хотя, если вы сейчас все поняли, то будет несложно.
 
 
-class Api {
-  // отлично что нет хардкода с url токеном, а данные передаются в конструктор.
+
+export class Api {
+
   constructor({ baseUrl, headers }) {
     this.baseUrl = baseUrl;
     this.headers = headers;
