+++ scripts/popup.js
@@ -1,5 +1,5 @@
 // создает объект попаппа, вешает слушатель на кнопку закрыть и метод открывашки
-class Popup {
+ export class Popup {
     constructor(popup) {
         this.popup = popup;
         this.popup.querySelector('.popup__close').addEventListener('click', this.close.bind(this));
@@ -15,7 +15,5 @@ class Popup {
 }
 
 
-
-// Нужно исправить: Каждому классу надо создавать отдельный файл, чтобы по названию было понятно что в этом файле лежит определённый класс
