+++ scripts/popupEx.js
@@ -1,5 +1,5 @@
 // спец попап для картинок
-class PopupImg extends Popup {
+export class PopupImg extends Popup {
   constructor(popup) {
     super(popup);
     this.img = this.popup.querySelector('.popup_img-op');
