+++ src/index.html
@@ -10,7 +10,7 @@
 <body>
   <div class="root">
     <header class="header root__section">
-      <img src="./images/logo.svg" alt="mesto logo" class="logo">
+      <img src="<%=require('./images/logo.svg')%>" alt="mesto logo" class="logo">
     </header>
     <div class="profile root__section">
       <div class="user-info">
@@ -29,7 +29,7 @@
     <!-- попап добавления карточек -->
     <div class="popup">
       <div class="popup__content">
-        <img src="./images/close.svg" alt="" class="popup__close">
+        <img src="<%=require('./images/close.svg')%>" alt="" class="popup__close">
         <h3 class="popup__title">Новое место</h3>
         <form class="popup__form" name="new" novalidate>
             <input type="text" required minlength="2" maxlength="30" name="name" class="popup__input popup__input_type_name" placeholder="Название">
@@ -43,7 +43,7 @@
     <!-- попап где редачить имя и фамилию -->
     <div class="popup popup_edit">
       <div class="popup__content">
-        <img src="./images/close.svg" alt="" class="popup__close">
+        <img src="<%=require('./images/close.svg')%>" alt="" class="popup__close">
         <h3 class="popup__title">Редактировать профиль</h3>
         <form class="popup__form" name="add">
             <input type="text" required minlength="2" maxlength="30" name="myName" class="popup__input popup__input_type_name" placeholder="Имя">
@@ -57,13 +57,13 @@
     <!-- попап широкоформатной картинки -->
     <div class="popup popup_img">
       <div class="popup_img-op">
-        <img src="./images/close.svg" alt="" class="popup__close">
+        <img src="<%=require('./images/close.svg')%>" alt="" class="popup__close">
       </div>
     </div>
     <!-- попап для смены аватарки -->
     <div class="popup popup_ava">
       <div class="popup__content">
-        <img src="./images/close.svg" alt="" class="popup__close">
+        <img src="<%=require('./images/close.svg')%>" alt="" class="popup__close">
         <h3 class="popup__title">Обновить аватар</h3>
         <form class="popup__form" name="ava">
             <input type="text" required minlength="2" name="avatar" class="popup__input popup__input_type_name" placeholder="URL">
