+++ webpack.config.js
@@ -2,6 +2,8 @@ const path = require('path');
 const MiniCssExtractPlugin = require("mini-css-extract-plugin");
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const WebpackMd5Hash = require('webpack-md5-hash');
+const OptimizeCssAssetsPlugin = require('optimize-css-assets-webpack-plugin');
+const isDev = process.env.NODE_ENV === 'development';
 
 module.exports = {
     entry: { main: './src/index.js' },
@@ -17,7 +19,7 @@ module: {
         },
         {
         test: /\.css$/,
-        use: [MiniCssExtractPlugin.loader, 'css-loader', 'postcss-loader']
+        use: [(isDev? 'style-loader' : MiniCssExtractPlugin.loader, 'css-loader', 'postcss-loader')]
         },
         {
         test: /\.(png|jpg|gif|ico|svg)$/,
@@ -28,6 +30,10 @@ module: {
                 options: {}
             },
             ]
+        },
+        {
+            test: /\.(eot|ttf|woff|woff2)$/,
+            loader: 'file-loader?name=./vendor/[name].[ext]'
         }
 
         ]
@@ -44,6 +50,17 @@ plugins: [
     new WebpackMd5Hash(),
     new webpack.DefinePlugin({
         'NODE_ENV': JSON.stringify(process.env.NODE_ENV)
-    })
+    }),
+    new MiniCssExtractPlugin({
+        filename: 'style.[contenthash].css'
+    }),
+    new OptimizeCssAssetsPlugin({
+        assetNameRegExp: /\.css$/g,
+        cssProcessor: require('cssnano'),
+        cssProcessorPluginOptions: {
+                preset: ['default'],
+        },
+        canPrint: true
+    }) 
 ]
 };
